requirements.txt
python-telegram-bot[webhooks]==20.8
SQLAlchemy==2.0.34
alembic==1.13.2
python-dotenv==1.0.1
python-dateutil==2.9.0
psycopg2-binary==2.9.9
pandas==2.2.3
matplotlib==3.9.2
openpyxl==3.1.5
Flask==3.0.3
gunicorn==22.0.0
PyJWT==2.9.0
google-generativeai>=0.8.0
Pillow>=10.0.0
pytest
pytest-mock
pytest-asyncio
